requirements.txt
langchain
openai
huggingface_hub
streamlit
python-dotenv
flask
django